requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiostream==0.5.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
bleach==6.1.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Modals==0.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.2.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
modal==0.64.127
multidict==6.1.0
pillow==10.4.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
rich==13.8.1
shellingham==1.5.4
sigtools==4.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
starlette==0.38.5
synchronicity==0.7.6
toml==0.10.2
typer==0.12.5
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.12.2
visitor==0.1.3
watchfiles==0.24.0
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.1
yarl==1.11.1
zipp==3.20.0
